Move BlockWithHead to separate file for tf as well
