+++ nvtabular/framework_utils/tensorflow/__init__.py
@@ -19,7 +19,8 @@ from .feature_column_utils import make_feature_column_workflow
 from .features import (FilterFeatures, ConcatFeatures, StackFeatures, TabularLayer, AsSparseLayer,
                        AsDenseLayer, ParseTokenizedText, AsTabular)
 from .layers.inputs import EmbeddingsLayer, TransformersTextEmbedding, InputFeatures
-from .blocks.base import right_shift_layer, Block, BlockWithHead, TabularBlock, SequentialBlock
+from .blocks.base import right_shift_layer, Block, TabularBlock, SequentialBlock
+from .blocks.with_head import BlockWithHead
 from .blocks.dlrm import DLRMBlock
 from ...loader.tensorflow import KerasSequenceLoader
 from .heads import Head, Task
